backend_v2/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
openai==1.45.0
langchain-openai==0.2.0
httpx==0.27.2
python-multipart==0.0.9
pandas
openpyxl # excel 지원
pydantic[email]

# uvicorn app.main:app --reload --port 8000
